istio: add kargo promoter
The cert manager kargo pipeline works, but it's not re-usable and is
heavily tied to the CertManager struct at the top level.

This patch creates a re-usable addon-promoter component that can promote
any add-on.  It's purpose is to configure a kargo promotion project and
pipeline stages to watch a helm chart and update a single field in a
yaml file.  This approach is generalized and reusable.