istio: refactor to import config data and embed yaml
This patch moves the top level Istio config struct from the root of the
repository into package istio field `Config`.  This is an effort to
refactor the large number of top level files so the example isn't so
overwhelming.

Istio configuration depends on the Organization configuration, so the
ProjectBuilder takes a platform.#Organization value as an input field
and builds the platform.#Project from it, accessible in the
PlatformBuilder.Platform field.

This is looking to be a common pattern and resembles good dependency
injection practices in other languages.

This patch also loads the istio version from a plain YAML file using the
cue embed functionality.