cert-manager: refactor to import config data and embed yaml
Similar to istio, this patch configures cert-manager to use the version
embedded in config/certmanager/cert-manager.yaml

The CUE embed functionality replaces the Holos instance data loading
functionality.  Kargo is reconfigured to edit the yaml file embedded in
the cue package.

The CertManager top level struct has been replaced, moved to package
certmanager.Config accessible through an import.

This patch doesn't move the project components out of the main package,
so we don't yet need a certmanager.ProjectBuilder.  cert-manager is part
of the security project, so it'll probably end up in that package
anyway.