unity: add httproutes test case for unity
Previously we didn't have a good test case for the httproutes slow down
issue I've run into a couple of times.  The issue is related to the for
loop we use to mix backend references into the HTTPRoute resource.

This patch adds a test case with minimal external dependencies.  We do
not execute any sub processes (e.g. helm) by using the holos show
buildplans command to produce the complete build plan from a single Go
process.

We use a label selector to focus on the Holos httproutes component
experiencing the problem.

This patch assumes the CUE folks are using the `scripts/unity`
executable to execute unity tests.