refactor the platform configuration domain model into packages
This patch begins refactoring the domain model into packages.  The goal
is to establish a platform package with types in pkg/types/platform and
concrete configuration values in pkg/config/platform.

The platform spec is in the main package, also named platform in
./platform

Components will import the main platform package from
example.com/holos/platform to get a handle on aspects of the Platform
spec.

The main platform package imports the platform configuration from
example.com/holos/pkg/config/platform, iterating over platform.Stacks to
assemble the Platform.spec.components field.

It may be worth defining a #PlatformBuilder which takes in parameters
and assembles a Platform resource, but we'll see how this comes
together.

It may also be worth separating out the #FooBuilder definitions from the
types, the way things are assembled is more specific to the organization
than the definitions of the things being assembled.

Result:

At the end of this change list, the deploy directory should be able to
be removed and re-created as it was prior with `holos render platform`.
This is intended to be a true refactor.

Once complete, we'll add additional functionality.  If additional
functionality is needed while refactoring, record it as a follow up
task.

The change list will work through components in the order listed in
scripts/apply