Wire up the bare argocd and security stacks
The argocd crds are the first component to manage.  Establish the argocd
Stack to manage the component.  We'll need to manage the argocd
namespace.

The security stack contains the namespaces component.  The current
behavior is for this component to manage all namespaces for all
components in a single BuildPlan.

We iterate over all stacks and manage the namespaces for each stack from
the namespaces field.  The Namespace kubernetes resource is largely
passed as-is all the way from the type definition.  This allows
composition at each step along the way.

Follow up work.

This patch relocates the namespaces component output directory to the
desired location of stacks/security/components/namespaces.  We'll remove
this in a follow up patch to preserve the integrity of the refactor,
currently components/namespaces.