Refactor podinfo kargo project (note: httproutes slows to 35s)
This patch is a major re-working of the previous KargoProjectBuilder to
a platform.#ProjectBuilder definition.  We define a "project" in the
context of a holos platform as a Kargo Project specifically.  It's a
special kind of Stack where a single input component definition is
expanded to multiple component build plans varying by the kargo project
promotion stages.

This could use some serious cleaning up and futher refactoring but it
does work and produces the same output manifests.

Note the httproutes component is much slower now, taking 35 seconds.
This patch leaves the slow down in so we can troubleshoot.  It seems
related to additional for loops used in the project builder.

Avoid httproutes with a selector.

    holos render platform --selector holos.run/component.name!=httproutes