Refactor istio-base component in network stack
Istio is effectively a sub-stack in and of itself, but there's no need
to define a separate collection for it.  Configuration values used by
multiple components are imported from example.com/holos/pkg/config/istio
and accessible as istio.config

The istio Kargo Project will be handled separately, but the istio config
package should support this use case just fine.