Refactor istio-promoter shared component in network stack
The addon-promoter component is intended to be shared by multiple
components in multiple stacks.  It lives next to namespaces in
stacks/shared/components as a result.

The component is configured primarily through input parameters passed
from the Platform spec.  Shared configuration values for istio are in
the istio config package.