Handle multiple credentials during reset
Previously the reset-cluster script was hard-coded to check for kargo
credentials.  Other examples may use other credentials, so the script
changes to apply all manifests in $(mkcert -CAROOT)/manifests/
immediately after a reset.

The kargo-git-creds similarly changes to write the credentials to a file
unique to the github app installation id so it doesn't write over other
manifests in the manifests directory.