Fix argocd redis secret not reconciling
There's a race between the argocd controller and redis against the auth
password.  When the cluster is reset, the argocd web ui often returns
401 NOAUTH errors from redis.

This patch manages the redis auth secret prior to managing the ArgoCD
component.  This avoids the race, both redis and argocd come up
configured with the correct password.