Define #PromoterBuilder to mix add-on promoters into stacks
Previously each add-on promoter configuration was declared as a separate
component.  This patch defines #PromoterBuilder which partially builds
the fields for a #StackBuilder to compose in an *-promoter component for
an add-on component.

The purpose is to establish one definition of the configuration of how
add-ons updates are promoted to production.