manage external secrets on multiple clusters
For the Progressive Delivery topic, we want to show how holos makes it
easy to mix in promotion stages to platform components such as the
external-secrets operator.

This patch established the holos.example/config/platform.clusters struct
which is a collection of all clusters composing the platform.  Platform
components to be managed on each cluster are defined in the
platform.clusters.components field.

Adding and removing files to config/platform/clusters/*.yaml are the
onlything necessary to add and remove clusters to the platform.

The standard list of components is defined by
platform.#ClusterComponents

Individual cluster yaml files may override software versions, or any
component parameter using the respective versions and components fields.